backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==5.0.0
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==4.2.16
django-cachalot==2.6.3
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-ckeditor==6.7.1
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==24.3
django-js-asset==2.2.0
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.11.1
flower==2.0.1
gunicorn==23.0.0
humanize==4.11.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
msgpack==1.1.0
packaging==24.1
pillow==11.0.0
pip==21.2.4
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.1
setuptools==58.0.4
six==1.16.0
sqlparse==0.5.1
stripe==11.2.0
text-unidecode==1.3
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.15
vine==5.1.0
wcwidth==0.2.13
drf-spectacular
drf-spectacular-sidecar

